# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  address: string;
  chain: string;
  category: string;
  label: string;
  /** @nullable */
  notes?: string | null;
  isActive: boolean;
  /** @nullable */
  lastActivity?: string | null;
  /** @nullable */
  totalVolume30d?: number | null;
  eventCount?: number;
  createdAt: string;
}

export interface WalletInput {
  /** @minLength 10 */
  address: string;
  chain: string;
  category: string;
  label: string;
  notes?: string;
}

export interface WalletUpdate {
  label?: string;
  notes?: string;
  category?: string;
  isActive?: boolean;
}

export interface Watchlist {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  walletCount: number;
  createdAt: string;
}

export interface WatchlistDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  wallets: Wallet[];
  createdAt: string;
}

export interface WatchlistInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface WatchlistUpdate {
  name?: string;
  description?: string;
}

export interface WatchlistWalletInput {
  walletId: number;
}

export interface OnchainEvent {
  id: number;
  walletId: number;
  walletAddress: string;
  walletLabel: string;
  eventType: string;
  chain: string;
  /** @nullable */
  tokenSymbol?: string | null;
  /** @nullable */
  amountUsd?: number | null;
  /** @nullable */
  txHash?: string | null;
  significance: string;
  /** @nullable */
  summary?: string | null;
  detectedAt: string;
}

export interface OnchainEventDetail {
  id: number;
  walletId: number;
  walletAddress: string;
  walletLabel: string;
  eventType: string;
  chain: string;
  /** @nullable */
  tokenSymbol?: string | null;
  /** @nullable */
  amountUsd?: number | null;
  /** @nullable */
  txHash?: string | null;
  significance: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  aiAnalysis?: string | null;
  /** @nullable */
  inferredIntent?: string | null;
  /** @nullable */
  historicalContext?: string | null;
  detectedAt: string;
}

export interface CountBreakdown {
  label: string;
  count: number;
}

export interface EventStats {
  totalEvents: number;
  last24h: number;
  last7d: number;
  byType: CountBreakdown[];
  byChain: CountBreakdown[];
  bySignificance: CountBreakdown[];
}

export interface Alert {
  id: number;
  name?: string;
  channel: string;
  /** @nullable */
  webhookUrl?: string | null;
  /** @nullable */
  chatId?: string | null;
  minSignificance?: string;
  isActive: boolean;
  createdAt: string;
}

export interface AlertInput {
  name: string;
  channel: string;
  webhookUrl?: string;
  chatId?: string;
  minSignificance?: string;
}

export interface AlertUpdate {
  name?: string;
  webhookUrl?: string;
  chatId?: string;
  minSignificance?: string;
  isActive?: boolean;
}

export interface IntelligenceItem {
  id: number;
  headline: string;
  body: string;
  category: string;
  significance: string;
  relatedWallets?: string[];
  /** @nullable */
  relatedEventId?: number | null;
  createdAt: string;
}

export interface AnalysisRequest {
  walletId?: number;
  eventId?: number;
  prompt?: string;
}

export interface AnalysisResult {
  analysis: string;
  inferredIntent: string;
  significance: string;
  historicalContext?: string;
  confidence?: string;
}

export interface MarketSummary {
  headline: string;
  body: string;
  trendingNarratives: string[];
  topActiveChains: string[];
  generatedAt: string;
}

export interface DashboardStats {
  trackedWallets: number;
  activeWatchlists: number;
  eventsToday: number;
  criticalEvents: number;
  totalVolumeTracked: number;
  activeAlerts: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  significance: string;
  /** @nullable */
  chain?: string | null;
  /** @nullable */
  walletLabel?: string | null;
  /** @nullable */
  amountUsd?: number | null;
  timestamp: string;
}

export type ListWalletsParams = {
watchlistId?: number;
category?: string;
chain?: string;
};

export type ListEventsParams = {
walletId?: number;
eventType?: string;
chain?: string;
significance?: string;
limit?: number;
};

export type GetIntelligenceFeedParams = {
limit?: number;
};

export type GetDashboardActivityParams = {
limit?: number;
};