# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Argus — AI Onchain Intelligence & Whale Tracking Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List detected onchain events
      parameters:
        - name: walletId
          in: query
          schema:
            type: integer
        - name: eventType
          in: query
          schema:
            type: string
        - name: chain
          in: query
          schema:
            type: string
        - name: significance
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnchainEvent"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get single event with full AI analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event with analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnchainEventDetail"
        "404":
          description: Not found

  # ── Alerts ────────────────────────────────────────────────────────────────
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alert configurations
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create alert configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete alert configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Intelligence ──────────────────────────────────────────────────────────
  /intelligence/feed:
    get:
      operationId: getIntelligenceFeed
      tags: [intelligence]
      summary: Get AI-generated intelligence summaries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Intelligence feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntelligenceItem"
  /intelligence/analyze:
    post:
      operationId: analyzeWallet
      tags: [intelligence]
      summary: Trigger AI analysis for a wallet or event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
...
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Recent activity feed for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    OnchainEvent:
      type: object
      required: [id, walletId, walletAddress, walletLabel, eventType, chain, significance, detectedAt]
      properties:
        id:             { type: integer }
        walletId:       { type: integer }
        walletAddress:  { type: string }
        walletLabel:    { type: string }
        eventType:      { type: string }
        chain:          { type: string }
        tokenSymbol:    { type: ["string", "null"] }
        amountUsd:      { type: ["number", "null"] }
        txHash:         { type: ["string", "null"] }
        significance:   { type: string }
        summary:        { type: ["string", "null"] }
        detectedAt:     { type: string }
    OnchainEventDetail:
      type: object
      required: [id, walletId, walletAddress, walletLabel, eventType, chain, significance, detectedAt]
      properties:
        id:             { type: integer }
        walletId:       { type: integer }
        walletAddress:  { type: string }
        walletLabel:    { type: string }
        eventType:      { type: string }
        chain:          { type: string }
        tokenSymbol:    { type: ["string", "null"] }
        amountUsd:      { type: ["number", "null"] }
        txHash:         { type: ["string", "null"] }
        significance:   { type: string }
        summary:        { type: ["string", "null"] }
        aiAnalysis:     { type: ["string", "null"] }
        inferredIntent: { type: ["string", "null"] }
        historicalContext: { type: ["string", "null"] }
        detectedAt:     { type: string }
...
    Alert:
      type: object
      required: [id, channel, isActive, createdAt]
      properties:
        id:         { type: integer }
        name:       { type: string }
        channel:    { type: string }
        webhookUrl: { type: ["string", "null"] }
        chatId:     { type: ["string", "null"] }
        minSignificance: { type: string }
        isActive:   { type: boolean }
        createdAt:  { type: string }
    AlertInput:
      type: object
      required: [channel, name]
      properties:
        name:       { type: string }
        channel:    { type: string }
        webhookUrl: { type: string }
        chatId:     { type: string }
        minSignificance: { type: string }
    AlertUpdate:
      type: object
      properties:
        name:       { type: string }
        webhookUrl: { type: string }
        chatId:     { type: string }
        minSignificance: { type: string }
        isActive:   { type: boolean }

    # ── Intelligence ─────────────────────────────────────────────────────────
    IntelligenceItem:
      type: object
      required: [id, headline, body, category, significance, createdAt]
      properties:
        id:          { type: integer }
        headline:    { type: string }
        body:        { type: string }
        category:    { type: string }
        significance: { type: string }
        relatedWallets:
          type: array
          items:
            type: string
        relatedEventId: { type: ["integer", "null"] }
        createdAt:   { type: string }
    AnalysisRequest:
      type: object
      properties:
        walletId:  { type: integer }
        eventId:   { type: integer }
        prompt:    { type: string }
    AnalysisResult:
      type: object
      required: [analysis, inferredIntent, significance]
      properties:
        analysis:        { type: string }
        inferredIntent:  { type: string }
        significance:    { type: string }
        historicalContext: { type: string }
        confidence:      { type: string }
...
    ActivityItem:
      type: object
      required: [id, type, title, description, significance, timestamp]
      properties:
        id:          { type: integer }
        type:        { type: string }
        title:       { type: string }
        description: { type: string }
        significance: { type: string }
        chain:       { type: ["string", "null"] }
        walletLabel: { type: ["string", "null"] }
        amountUsd:   { type: ["number", "null"] }
        timestamp:   { type: string }