# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  significance: string;
  /** @nullable */
  chain?: string | null;
  /** @nullable */
  walletLabel?: string | null;
  /** @nullable */
  amountUsd?: number | null;
  timestamp: string;
}