# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  name?: string;
  channel: string;
  /** @nullable */
  webhookUrl?: string | null;
  /** @nullable */
  chatId?: string | null;
  minSignificance?: string;
  isActive: boolean;
  createdAt: string;
}