# lib/api-zod/src/generated/types/intelligenceItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface IntelligenceItem {
  id: number;
  headline: string;
  body: string;
  category: string;
  significance: string;
  relatedWallets?: string[];
  /** @nullable */
  relatedEventId?: number | null;
  createdAt: string;
}