# lib/api-zod/src/generated/types/onchainEventDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface OnchainEventDetail {
  id: number;
  walletId: number;
  walletAddress: string;
  walletLabel: string;
  eventType: string;
  chain: string;
  /** @nullable */
  tokenSymbol?: string | null;
  /** @nullable */
  amountUsd?: number | null;
  /** @nullable */
  txHash?: string | null;
  significance: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  aiAnalysis?: string | null;
  /** @nullable */
  inferredIntent?: string | null;
  /** @nullable */
  historicalContext?: string | null;
  detectedAt: string;
}